refactor: extract penca section component
